c++ store practice
This is a program in which I created on aug 12 to enhance my knowledge and skills regarding about the usage of c++....
This program has the ability to store value and change value of the products' prices. 
It also has the ability to display and return right at the start as you wish it to.
Mind that this is an unfinished "fun little project" of mine.

You have my permission to copy it if you desire to.